+++ tests/testthat/test_lm.R
@@ -17,10 +17,10 @@ context("lm-parsnip")
 
 lm_parsnip <- function(...) {
   parsnip::fit(
-    parsnip::set_engine(parsnip::linear_reg(), "lm"), 
+    parsnip::set_engine(parsnip::linear_reg(), "lm"),
     ...
   )
-} 
+}
 
 test_that("Predictions within threshold and parsed model results are equal", {
   expect_false(has_alert(lm_parsnip(mpg ~ wt, offset = am, data = df)))
@@ -36,5 +36,13 @@ test_that("Model can be saved and re-loaded", {
   mp <- tempfile(fileext = ".yml")
   yaml::write_yaml(parse_model(model), mp)
   l <- yaml::read_yaml(mp)
-  expect_silent(tidypredict_fit(l))
+  pm <- as_parsed_model(l)
+  expect_silent(tidypredict_fit(pm))
+})
+
+test_that("tidy() works", {
+  expect_is(
+    tidy(parse_model(lm(mpg ~., mtcars))),
+    "tbl_df"
+  )  
 })