+++ tests/testthat/test_xgboost.R
@@ -40,10 +40,17 @@ test_that("Confirm SQL function returns a query", {
   )
 })
 
+context("xgboost-parsnip")
+test_that("Predictions are correct for different objectives", {
+  m <- parsnip::fit(parsnip::set_engine(parsnip::boost_tree(), "xgboost"), am ~ ., data = mtcars)
+  expect_false(tidypredict_test(m, df = mtcars)$alert)
+})
+
 context("xgboost-saved")
 test_that("Model can be saved and re-loaded", {
   mp <- tempfile(fileext = ".yml")
   yaml::write_yaml(parse_model(xgb_reglinear), mp)
   l <- yaml::read_yaml(mp)
-  expect_silent(tidypredict_fit(l))
+  pm <- as_parsed_model(l)
+  expect_silent(tidypredict_fit(pm))
 })